Revert "Issue 1970"
