+++ src/abilities/Vehemoth.js
@@ -3,7 +3,7 @@ import { Team } from '../utility/team';
 import * as matrices from '../utility/matrices';
 import * as arrayUtils from '../utility/arrayUtils';
 import { Effect } from '../effect';
-import { Direction } from '../frontend/hex';
+import { Direction } from '../utility/hex';
 
 /** Creates the abilities
  * @param {Game} G the game object
