+++ src/creature.js
@@ -1,7 +1,7 @@
 import * as $j from 'jquery';
 import { Ability } from './ability';
 import { search } from './utility/pathfinding';
-import { PhaserHex } from './frontend/phaser/phaser_hex';
+import { Hex } from './utility/hex';
 import * as arrayUtils from './utility/arrayUtils';
 import { Drop } from './drops';
 
@@ -809,12 +809,11 @@ export class Creature {
 				return; // Break if empty path
 			}
 
-			game.grid.xray(new PhaserHex(0, 0, false, game)); // Clean Xray
+			game.grid.xray(new Hex(0, 0, false, game)); // Clean Xray
 
 			this.travelDist = 0;
 
-			game.animations.callMethodByStr(opts.animation, this, path, opts);
-			// game.animations[opts.animation](this, path, opts);
+			game.animations[opts.animation](this, path, opts);
 		} else {
 			game.log('This creature cannot be moved');
 		}
