+++ src/utility/arrayUtils.ts
@@ -1,6 +1,6 @@
-import { Creature } from '../frontend/creature';
-import { Direction, Hex } from '../frontend/hex';
-import { HexGrid } from '../frontend/hexgrid';
+import { Creature } from '../creature';
+import { Direction, Hex } from './hex';
+import { HexGrid } from './hexgrid';
 
 /**
  * Find an object in the current Array based on its pos attribute.
