+++ src/utility/hexgrid.ts
@@ -1,13 +1,13 @@
+import * as $j from 'jquery';
 import { Direction, Hex } from './hex';
-import { Creature } from './creature';
-import { search } from '../utility/pathfinding';
-import * as matrices from '../utility/matrices';
-import { Team, isTeam } from '../utility/team';
+import { Creature } from '../creature';
+import { search } from './pathfinding';
+import * as matrices from './matrices';
+import { Team, isTeam } from './team';
+import * as arrayUtils from './arrayUtils';
 import Game from '../game';
 import { Trap } from './trap';
-import * as arrayUtils from '../utility/arrayUtils';
-
-export interface QueryOptions {
+interface QueryOptions {
 	/**
 	 * Target team.
 	 */
@@ -87,9 +87,8 @@ export interface QueryOptions {
  * Object containing grid and methods concerning the whole grid.
  * Should only have one instance during the game.
  */
-export abstract class HexGrid {
+export class HexGrid {
 	game: Game;
-	opts: any; // Creation options.
 
 	/**
 	 * Contain all hexes in row arrays (hexes[y][x]).
@@ -104,22 +103,33 @@ export abstract class HexGrid {
 	 */
 	lastClickedHex: Hex;
 
+	display: Phaser.Sprite;
+	gridGroup: Phaser.Group;
+	trapGroup: Phaser.Group;
+	hexesGroup: Phaser.Group;
+	displayHexesGroup: Phaser.Group;
+	overlayHexesGroup: Phaser.Group;
+	inputHexesGroup: Phaser.Group;
+	dropGroup: Phaser.Group;
+	creatureGroup: Phaser.Group;
+	trapOverGroup: Phaser.Group;
 	selectedHex: Hex;
 	_executionMode: boolean;
+	materialize_overlay: any;
 	lastQueryOpt: any;
 
 	/* Constructor
 	 *
 	 * Create attributes and populate JS grid with Hex objects
 	 */
-	constructor(opts: any, game: Game) {
+	constructor(opts, game: Game) {
 		const defaultOpt = {
 			nbrRow: 9,
 			nbrhexesPerRow: 16,
 			firstRowFull: false,
 		};
 
-		this.opts = { ...defaultOpt, ...opts };
+		opts = { ...defaultOpt, ...opts };
 
 		this.game = game;
 		this.hexes = []; // Hex Array
@@ -127,46 +137,95 @@ export abstract class HexGrid {
 		this.allhexes = []; // All hexes
 		this.lastClickedHex = undefined;
 
-		// If true, clicking a monster will instantly kill it.
-		this._executionMode = false;
-	}
+		this.display = game.Phaser.add.group(undefined, 'displayGroup');
+		this.display.x = 230;
+		this.display.y = 380;
 
-	/**
-	 * Populates hexes, allhexes and selectedHex.
-	 */
-	populate() {
+		this.gridGroup = game.Phaser.add.group(this.display, 'gridGroup');
+		this.gridGroup.scale.set(1, 0.75);
+
+		this.trapGroup = game.Phaser.add.group(this.gridGroup, 'trapGrp');
+		this.hexesGroup = game.Phaser.add.group(this.gridGroup, 'hexesGroup');
+		this.displayHexesGroup = game.Phaser.add.group(this.gridGroup, 'displayHexesGroup');
+		this.overlayHexesGroup = game.Phaser.add.group(this.gridGroup, 'overlayHexesGroup');
+		this.inputHexesGroup = game.Phaser.add.group(this.gridGroup, 'inputHexesGroup');
+		this.dropGroup = game.Phaser.add.group(this.display, 'dropGrp');
+		this.creatureGroup = game.Phaser.add.group(this.display, 'creaturesGrp');
+		// Parts of traps displayed over creatures
+		this.trapOverGroup = game.Phaser.add.group(this.display, 'trapOverGrp');
+		this.trapOverGroup.scale.set(1, 0.75);
+
 		// Populate grid
-		for (let row = 0; row < this.opts.nbrRow; row++) {
+		for (let row = 0; row < opts.nbrRow; row++) {
 			this.hexes.push([]);
-			for (let hex = 0, len = this.opts.nbrhexesPerRow; hex < len; hex++) {
-				if (hex == this.opts.nbrhexesPerRow - 1) {
-					if (
-						(row % 2 == 0 && !this.opts.firstRowFull) ||
-						(row % 2 == 1 && this.opts.firstRowFull)
-					) {
+			for (let hex = 0, len = opts.nbrhexesPerRow; hex < len; hex++) {
+				if (hex == opts.nbrhexesPerRow - 1) {
+					if ((row % 2 == 0 && !opts.firstRowFull) || (row % 2 == 1 && opts.firstRowFull)) {
 						continue;
 					}
 				}
 
-				this.hexes[row][hex] = this.createHex(hex, row, this, this.game);
+				this.hexes[row][hex] = new Hex(hex, row, this);
 				this.allhexes.push(this.hexes[row][hex]);
 			}
 		}
 
 		this.selectedHex = this.hexes[0][0];
+
+		// If true, clicking a monster will instantly kill it.
+		this._executionMode = false;
+
+		// Events
+		this.game.signals.metaPowers.add(this.handleMetaPowerEvent, this);
 	}
 
-	/**
-	 * Create a implementation Hex object.
-	 *
-	 * @param x
-	 * @param y
-	 * @param grid
-	 * @param game
-	 */
-	abstract createHex(x: number, y: number, grid: HexGrid, game?: Game): Hex;
+	handleMetaPowerEvent(message, payload) {
+		if (message === 'toggleExecuteMonster') {
+			this._executionMode = payload;
+		}
+	}
 
-	abstract querySelf(o: any);
+	querySelf(o) {
+		const game = this.game;
+		const defaultOpt = {
+			fnOnConfirm: () => {
+				// No-op function.
+			},
+			fnOnSelect: (creature: Creature) => {
+				creature.hexagons.forEach((hex) => {
+					hex.overlayVisualState('creature selected player' + hex.creature.team);
+				});
+			},
+			fnOnCancel: () => {
+				this.game.activeCreature.queryMove();
+			},
+			args: {},
+			confirmText: 'Confirm',
+			id: game.activeCreature.id,
+		};
+
+		o = { ...defaultOpt, ...o };
+
+		game.activeCreature.hint(o.confirmText, 'confirm');
+
+		this.queryHexes({
+			fnOnConfirm: (hex, args) => {
+				args.opt.fnOnConfirm(game.activeCreature, args.opt.args, { queryOptions: o });
+			},
+			fnOnSelect: (hex, args) => {
+				args.opt.fnOnSelect(game.activeCreature, args.opt.args);
+			},
+			fnOnCancel: (hex, args) => {
+				args.opt.fnOnCancel(game.activeCreature, args.opt.args);
+			},
+			args: {
+				opt: o,
+			},
+			hexes: game.activeCreature.hexagons,
+			hideNonTarget: true,
+			id: o.id,
+		});
+	}
 
 	/**
 	 * Shortcut to queryChoice with specific directions.
@@ -506,7 +565,113 @@ export abstract class HexGrid {
 	 * @param {boolean} o.replaceEmptyHexesWithDashed Replace all non targetable, empty hexes with dashed hexes.
 	 * 	o.hexesDashed will override this option.
 	 */
-	abstract queryCreature(o);
+	queryCreature(o) {
+		const game = this.game;
+		const defaultOpt = {
+			fnOnConfirm: () => {
+				game.activeCreature.queryMove();
+			},
+			fnOnSelect: (creature) => {
+				creature.tracePosition({
+					overlayClass: 'creature selected player' + creature.team,
+				});
+			},
+			fnOnCancel: () => {
+				game.activeCreature.queryMove();
+			},
+			optTest: () => true,
+			args: {},
+			hexes: [],
+			hexesDashed: [],
+			flipped: false,
+			id: 0,
+			team: Team.Enemy,
+			replaceEmptyHexesWithDashed: false,
+		};
+
+		o = { ...defaultOpt, ...o };
+
+		/* Divide hexes into:
+		- containing valid targets
+		- empty (no possible target)
+		Hexes containing invalid targets (wrong team, o.optTest, etc) are discard. */
+		const { targetHexes, emptyHexes } = o.hexes.reduce(
+			(acc, hex) => {
+				const sourceCreature = game.creatures[o.id];
+				const targetCreature = hex.creature;
+
+				const acceptTargetHex = () => {
+					return {
+						...acc,
+						targetHexes: [...acc.targetHexes, hex],
+					};
+				};
+
+				const acceptEmptyHex = () => {
+					return {
+						...acc,
+						emptyHexes: [...acc.emptyHexes, hex],
+					};
+				};
+
+				const discardHex = () => {
+					return acc;
+				};
+
+				if (!targetCreature) {
+					return acceptEmptyHex();
+				}
+
+				if (targetCreature instanceof Creature && targetCreature.id !== o.id) {
+					if (!o.optTest(hex.creature)) {
+						return discardHex();
+					}
+
+					if (isTeam(sourceCreature, targetCreature, o.team)) {
+						return acceptTargetHex();
+					}
+				}
+
+				return discardHex();
+			},
+			{ targetHexes: [], emptyHexes: [] },
+		);
+
+		o.hexes = targetHexes;
+
+		if (o.replaceEmptyHexesWithDashed && !o.hexesDashed.length) {
+			o.hexesDashed = emptyHexes;
+		}
+
+		let extended = [];
+		/* Add creature hexes that extend out of the range of the source hexes, so the
+		entire creature can be highlighted. */
+		o.hexes.forEach((hex) => {
+			extended = extended.concat(hex.creature.hexagons);
+		});
+
+		o.hexes = extended;
+
+		this.queryHexes({
+			fnOnConfirm: (hex, args) => {
+				const { creature } = hex;
+				args.opt.fnOnConfirm(creature, args.opt.args, { queryOptions: o });
+			},
+			fnOnSelect: (hex, args) => {
+				const { creature } = hex;
+				args.opt.fnOnSelect(creature, args.opt.args);
+			},
+			fnOnCancel: o.fnOnCancel,
+			args: {
+				opt: o,
+			},
+			hexes: o.hexes,
+			hexesDashed: o.hexesDashed,
+			flipped: o.flipped,
+			hideNonTarget: true,
+			id: o.id,
+		});
+	}
 
 	redoLastQuery() {
 		this.queryHexes(this.lastQueryOpt);
@@ -521,7 +686,301 @@ export abstract class HexGrid {
 	 * hexes : 		Array : 	Reachable hexes
 	 * callbackAfterQueryHexes : 		Function : 	empty function to be overridden with custom logic to execute after queryHexes
 	 */
-	abstract queryHexes(o);
+	queryHexes(o) {
+		const game = this.game;
+		const defaultOpt = {
+			fnOnConfirm: () => {
+				game.activeCreature.queryMove();
+			},
+			fnOnSelect: (hex: Hex) => {
+				game.activeCreature.faceHex(hex, undefined, true);
+				hex.overlayVisualState('creature selected player' + game.activeCreature.team);
+			},
+			fnOnCancel: () => {
+				game.activeCreature.queryMove();
+			},
+			callbackAfterQueryHexes: () => {
+				// empty function to be overridden with custom logic to execute after queryHexes
+			},
+			args: {},
+			hexes: [],
+			hexesDashed: [],
+			shrunkenHexes: [],
+			size: 1,
+			id: 0,
+			flipped: false,
+			hideNonTarget: false,
+			ownCreatureHexShade: false,
+			targeting: true,
+			fillHexOnHover: true,
+		};
+
+		o = { ...defaultOpt, ...o };
+
+		this.lastClickedHex = undefined;
+
+		// Save the last Query
+		this.lastQueryOpt = { ...o };
+
+		this.updateDisplay();
+		// Block all hexes
+		this.forEachHex((hex) => {
+			hex.unsetReachable();
+
+			if (o.hideNonTarget) {
+				hex.setNotTarget();
+			} else {
+				hex.unsetNotTarget();
+			}
+
+			if (o.hexesDashed.indexOf(hex) !== -1) {
+				hex.displayVisualState('dashed');
+			} else {
+				hex.cleanDisplayVisualState('dashed');
+			}
+
+			if (o.shrunkenHexes.includes(hex)) {
+				hex.displayVisualState('shrunken');
+			} else {
+				hex.cleanDisplayVisualState('shrunken');
+			}
+		});
+
+		// Cleanup
+		if (this.materialize_overlay) {
+			this.materialize_overlay.alpha = 0;
+		}
+
+		if (!o.ownCreatureHexShade) {
+			if (o.id instanceof Array) {
+				o.id.forEach((id) => {
+					game.creatures[id].hexagons.forEach((hex) => {
+						hex.overlayVisualState('ownCreatureHexShade');
+					});
+				});
+			} else {
+				if (o.id != 0) {
+					game.creatures[o.id].hexagons.forEach((hex) => {
+						hex.overlayVisualState('ownCreatureHexShade');
+					});
+				}
+			}
+		}
+
+		// Set reachable the given hexes
+		o.hexes.forEach((hex) => {
+			hex.setReachable();
+			if (o.hideNonTarget) {
+				hex.unsetNotTarget();
+			}
+			if (o.targeting) {
+				if (hex.creature instanceof Creature) {
+					if (hex.creature.id != this.game.activeCreature.id) {
+						hex.overlayVisualState('hover h_player' + hex.creature.team);
+					}
+				} else {
+					hex.overlayVisualState('hover h_player' + this.game.activeCreature.team);
+				}
+			}
+		});
+
+		if (o.callbackAfterQueryHexes) {
+			o.callbackAfterQueryHexes();
+		}
+
+		const onCreatureHover = (creature: Creature, queueEffect, hex: Hex) => {
+			if (creature.isDarkPriest()) {
+				if (creature === game.activeCreature) {
+					if (creature.hasCreaturePlayerGotPlasma()) {
+						creature.displayPlasmaShield();
+					}
+				} else {
+					creature.displayHealthStats();
+				}
+			}
+			creature.hexagons.forEach((h) => {
+				// Flashing outline
+				h.overlayVisualState('hover h_player' + creature.team);
+			});
+			if (creature !== game.activeCreature) {
+				if (!hex.reachable) {
+					$j('canvas').css('cursor', 'n-resize');
+				} else {
+					// Filled hex with color
+					hex.displayVisualState('creature player' + hex.creature.team);
+				}
+			}
+			queueEffect(creature.id);
+		};
+
+		// ONCLICK
+		const onConfirmFn = (hex: Hex) => {
+			// Debugger
+			const y = hex.y;
+			let x = hex.x;
+
+			// Clear display and overlay
+			$j('canvas').css('cursor', 'pointer');
+
+			if (this._executionMode && hex.creature instanceof Creature) {
+				hex.creature.die(
+					/* Target creature was killed by this fake "creature". This works because
+					the death logic doesn't actually care about the killing creature, just
+					that creature's player. The first player is always responsible for executing
+					creatures. */
+					{ player: game.players[0] },
+				);
+				return;
+			}
+
+			// Not reachable hex
+			if (!hex.reachable) {
+				this.lastClickedHex = undefined;
+
+				if (hex.creature instanceof Creature) {
+					// If creature
+					onCreatureHover(
+						hex.creature,
+						game.activeCreature !== hex.creature
+							? game.UI.bouncexrayQueue.bind(game.UI)
+							: game.UI.xrayQueue.bind(game.UI),
+						hex,
+					);
+				} else {
+					// If nothing
+					o.fnOnCancel(hex, o.args); // ON CANCEL
+				}
+			} else {
+				// Reachable hex
+				// Offset Pos
+				const offset = o.flipped ? o.size - 1 : 0;
+				const mult = o.flipped ? 1 : -1; // For flipped player
+
+				for (let i = 0, size = o.size; i < size; i++) {
+					// Try next hexagons to see if they fits
+					if (x + offset - i * mult >= this.hexes[y].length || x + offset - i * mult < 0) {
+						continue;
+					}
+
+					if (this.hexes[y][x + offset - i * mult].isWalkable(o.size, o.id)) {
+						x += offset - i * mult;
+						break;
+					}
+				}
+
+				hex = this.hexes[y][x]; // New coords
+				game.activeCreature.faceHex(hex, undefined, true, true);
+
+				if (hex !== this.lastClickedHex) {
+					this.lastClickedHex = hex;
+				}
+
+				o.fnOnConfirm(hex, o.args, { queryOptions: o });
+			}
+		};
+
+		const onHoverOffFn = (hex: Hex) => {
+			const { creature } = hex;
+
+			if (creature instanceof Creature) {
+				// toggle hover off event
+				if (creature.isDarkPriest()) {
+					// the plasma would have been displayed so now display the health again
+					creature.updateHealth();
+				}
+			}
+
+			$j('canvas').css('cursor', 'default');
+		};
+
+		// ONMOUSEOVER
+		const onSelectFn = (hex: Hex) => {
+			let { x } = hex;
+			const { y } = hex;
+
+			// Xray
+			this.xray(hex);
+
+			// Clear display and overlay
+			game.UI.xrayQueue(-1);
+			$j('canvas').css('cursor', 'pointer');
+
+			if (hex.creature instanceof Creature) {
+				// If creature
+				onCreatureHover(hex.creature, game.UI.xrayQueue.bind(game.UI), hex);
+			}
+
+			if (hex.reachable) {
+				if (o.fillHexOnHover) {
+					this.cleanHex(hex);
+					hex.displayVisualState('creature player' + this.game.activeCreature.team);
+				}
+
+				// Offset Pos
+				const offset = o.flipped ? o.size - 1 : 0;
+				const mult = o.flipped ? 1 : -1; // For flipped player
+
+				for (let i = 0, size = o.size; i < size; i++) {
+					// Try next hexagons to see if they fit
+					if (x + offset - i * mult >= this.hexes[y].length || x + offset - i * mult < 0) {
+						continue;
+					}
+
+					if (this.hexes[y][x + offset - i * mult].isWalkable(o.size, o.id)) {
+						x += offset - i * mult;
+						break;
+					}
+				}
+
+				hex = this.hexes[y][x]; // New coords
+				o.fnOnSelect(hex, o.args);
+			} else if (!hex.reachable) {
+				if (this.materialize_overlay) {
+					this.materialize_overlay.alpha = 0;
+				}
+				hex.overlayVisualState('hover');
+
+				$j('canvas').css('cursor', 'not-allowed');
+			}
+		};
+
+		// ONRIGHTCLICK
+		const onRightClickFn = (hex: Hex) => {
+			if (hex.creature instanceof Creature) {
+				game.UI.showCreature(hex.creature.type, hex.creature.player.id, 'grid');
+			} else {
+				if (game.activeCreature.isDarkPriest()) {
+					// If ability used, default to Dark Priest and say materialize has been used
+					if (game.activeCreature.abilities[3].used) {
+						game.UI.showCreature(
+							game.activeCreature.type,
+							game.activeCreature.player.id,
+							'emptyHex',
+						);
+					} else if (game.UI.lastViewedCreature) {
+						game.UI.showCreature(game.UI.lastViewedCreature, game.UI.selectedPlayer, 'emptyHex');
+					} else if (game.UI.selectedCreatureObj) {
+						game.UI.toggleDash(true);
+					} else {
+						game.UI.showCreature(
+							game.activeCreature.type,
+							game.activeCreature.player.id,
+							'emptyHex',
+						);
+					}
+				} else {
+					game.UI.showCreature(game.activeCreature.type, game.activeCreature.player.id, 'emptyHex');
+				}
+			}
+		};
+
+		this.forEachHex((hex) => {
+			hex.onSelectFn = onSelectFn;
+			hex.onHoverOffFn = onHoverOffFn;
+			hex.onConfirmFn = onConfirmFn;
+			hex.onRightClickFn = onRightClickFn;
+		});
+	}
 
 	/* xray(hex)
 	 *
@@ -543,7 +1002,7 @@ export abstract class HexGrid {
 				item.ghostOverlap();
 			});
 		} else {
-			// hex.ghostOverlap();
+			hex.ghostOverlap();
 		}
 	}
 
@@ -577,17 +1036,32 @@ export abstract class HexGrid {
 		}
 	}
 
-	abstract cleanHex(hex: Hex);
+	cleanHex(hex: Hex) {
+		hex.cleanDisplayVisualState();
+		hex.cleanOverlayVisualState();
+	}
 
-	/**
-	 * updateDisplay()
+	/* updateDisplay()
 	 *
 	 * Update overlay hexes with creature positions
 	 */
-	abstract updateDisplay();
+	updateDisplay() {
+		this.cleanDisplay();
+		this.cleanOverlay();
 
-	/**
-	 * hexExists(y, x)
+		this.hexes.forEach((hex) => {
+			hex.forEach((item) => {
+				if (item.creature instanceof Creature) {
+					if (item.creature.id == this.game.activeCreature.id) {
+						item.overlayVisualState(`active creature player${item.creature.team}`);
+						item.displayVisualState(`creature player${item.creature.team}`);
+					}
+				}
+			});
+		});
+	}
+
+	/* hexExists(y, x)
 	 *
 	 * x : 	Integer : 	Coordinates to test
 	 * y : 	Integer : 	Coordinates to test
@@ -605,8 +1079,7 @@ export abstract class HexGrid {
 		return false;
 	}
 
-	/**
-	 * isHexIn(hex, hexArray)
+	/* isHexIn(hex, hexArray)
 	 *
 	 * hex : 		Hex : 		Hex to look for
 	 * hexarray : 	Array : 	Array of hexes to look for hex in
@@ -638,7 +1111,7 @@ export abstract class HexGrid {
 		//	destination to test all hexagons
 		this.cleanReachable(); // If not pathfinding will bug
 		this.cleanPathAttr(true); // Erase all pathfinding data
-		search(this.hexes[y][x], this.createHex(-2, -2, null, this.game), size, id, this.game.grid);
+		search(this.hexes[y][x], new Hex(-2, -2, null, this.game), size, id, this.game.grid);
 
 		// Gather all the reachable hexes
 		const hexes: Hex[] = [];
@@ -666,7 +1139,7 @@ export abstract class HexGrid {
 		// Gather all the reachable hexes
 		let hexes = this.hexes[y][x].adjacentHex(distance);
 
-		hexes = hexes.filter((hex) => hex.isWalkable(size, id, true, false));
+		hexes = hexes.filter((hex) => hex.isWalkable(size, id, true));
 
 		return arrayUtils.extendToLeft(hexes, size, this.game.grid);
 	}
@@ -722,10 +1195,26 @@ export abstract class HexGrid {
 		return hexes;
 	}
 
-	abstract showGrid(val: any): void;
+	showGrid(val) {
+		this.forEachHex((hex) => {
+			if (hex.creature) {
+				hex.creature.xray(val);
+			}
+
+			if (hex.drop) {
+				return;
+			}
+
+			if (val) {
+				hex.displayVisualState('showGrid');
+			} else {
+				hex.cleanDisplayVisualState('showGrid');
+			}
+		});
+	}
 
 	// TODO: Rewrite methods used here to only require the creature as an argument.
-	showMovementRange(id: number): void {
+	showMovementRange(id) {
 		const creature = this.game.creatures[id];
 		const hexes = this.findCreatureMovementHexes(creature);
 
@@ -740,9 +1229,14 @@ export abstract class HexGrid {
 		});
 	}
 
-	abstract showCurrentCreatureMovementInOverlay(creature: Creature): void;
+	showCurrentCreatureMovementInOverlay(creature) {
+		//lastQueryOpt is same thing as used in redoQuery
+		this.lastQueryOpt.hexes.forEach((hex) => {
+			hex.overlayVisualState('hover h_player' + creature.team);
+		});
+	}
 
-	findCreatureMovementHexes(creature: Creature): Hex[] {
+	findCreatureMovementHexes(creature) {
 		if (creature.movementType() === 'flying') {
 			return this.getFlyingRange(
 				creature.x,
@@ -841,12 +1335,29 @@ export abstract class HexGrid {
 
 		this.redoLastQuery();
 		// Clear Xray.
-		this.xray(this.createHex(-1, -1, null, this.game));
+		this.xray(new Hex(-1, -1, null, this.game));
 		// Clear Xray Queue.
 		this.game.UI.xrayQueue(-1);
 	}
 
-	abstract orderCreatureZ();
+	orderCreatureZ() {
+		let index = 0;
+		const creatures = this.game.creatures;
+
+		for (let y = 0, leny = this.hexes.length; y < leny; y++) {
+			for (let i = 1, len = creatures.length; i < len; i++) {
+				if (creatures[i].y == y) {
+					this.creatureGroup.remove(creatures[i].grp);
+					this.creatureGroup.addAt(creatures[i].grp, index++);
+				}
+			}
+
+			if (this.materialize_overlay && this.materialize_overlay.posy == y) {
+				this.creatureGroup.remove(this.materialize_overlay);
+				this.creatureGroup.addAt(this.materialize_overlay, index++);
+			}
+		}
+	}
 
 	//******************//
 	//Shortcut functions//
@@ -890,6 +1401,24 @@ export abstract class HexGrid {
 		});
 	}
 
+	/* cleanDisplay(cssClass)
+	 *
+	 * cssClass : 	String : 	Class(es) name(s) to remove with jQuery removeClass function
+	 *
+	 * Shorcut for $allDispHex.removeClass()
+	 */
+	cleanDisplay(cssClass = '') {
+		this.forEachHex((hex) => {
+			hex.cleanDisplayVisualState(cssClass);
+		});
+	}
+
+	cleanOverlay(cssClass = '') {
+		this.forEachHex((hex) => {
+			hex.cleanOverlayVisualState(cssClass);
+		});
+	}
+
 	/* previewCreature(creatureData)
 	 *
 	 * pos : 			Object : 	Coordinates {x,y}
@@ -897,15 +1426,58 @@ export abstract class HexGrid {
 	 *
 	 * Draw a preview of the creature at the given coordinates
 	 */
-	abstract previewCreature(pos, creatureData, player);
+	previewCreature(pos, creatureData, player) {
+		const game = this.game;
+		const hex = this.hexes[pos.y][pos.x - (creatureData.size - 1)];
+
+		if (!this.materialize_overlay) {
+			// If sprite does not exists
+			// Adding sprite
+			this.materialize_overlay = this.creatureGroup.create(0, 0, creatureData.name + '_cardboard');
+			this.materialize_overlay.anchor.setTo(0.5, 1);
+			this.materialize_overlay.posy = pos.y;
+		} else {
+			this.materialize_overlay.loadTexture(creatureData.name + '_cardboard');
+			if (this.materialize_overlay.posy != pos.y) {
+				this.materialize_overlay.posy = pos.y;
+				this.orderCreatureZ();
+			}
+		}
+
+		// Placing sprite
+		this.materialize_overlay.x =
+			hex.displayPos.x +
+			(!player.flipped
+				? creatureData.display['offset-x']
+				: 90 * creatureData.size -
+				  this.materialize_overlay.texture.width -
+				  creatureData.display['offset-x']) +
+			this.materialize_overlay.texture.width / 2;
+		this.materialize_overlay.y =
+			hex.displayPos.y + creatureData.display['offset-y'] + this.materialize_overlay.texture.height;
+		this.materialize_overlay.alpha = 0.5;
+
+		if (player.flipped) {
+			this.materialize_overlay.scale.setTo(-1, 1);
+		} else {
+			this.materialize_overlay.scale.setTo(1, 1);
+		}
+
+		for (let i = 0, size = creatureData.size; i < size; i++) {
+			const hexInstance = this.hexes[pos.y][pos.x - i];
+			this.cleanHex(hexInstance);
+			hexInstance.overlayVisualState('creature selected player' + game.activeCreature.team);
+		}
+	}
 
 	/**
-	 * Internal debugging method to log
+	 * Internal debugging method to log and visually highlight (in blue) an array
 	 * of hexes.
 	 *
-	 * @param hexes Hexes to log.
+	 * @param hexes Hexes to log and visually highlight.
 	 */
 	__debugHexes(hexes: Hex[]) {
 		console.debug({ hexes }, hexes.map((hex) => hex.coord).join(', '));
+		hexes.forEach((hex) => hex.displayVisualState('creature selected player1'));
 	}
 }
